Add files via upload
Here you will build a simple console application in the browser theres a lot going on if you never used these concepts before. Also, if your taking CSC 374 for DePaul for the harder path you will want to see the notes I created following them and building on the Data structs from my other repo your going to pass with a high A I made sure to post what functions are doing to and best practice you might even enprove on some of the code used for the course helping promote better learning of Networking and server implementations.