Add files via upload
Hey I made the Dir inside this chapter file due to the learning process i found works base for this type of learning. I see it as progression based. FYI i mastered networking in a OOP lang that did a lot of dynamic linking and was interpreted so if you haven't learned the basics of networking with the trainning wheels on I hope the resources I provide helps a lot. All programs work and all explanations have been checked.